LayoutTests/userscripts/script-not-run-for-fragments-expected.txt

CONSOLE MESSAGE: line 1: injected: blank.xhtml
This test should not infinitely inject user scripts. 
